+++ CHANGELOG.md
@@ -8,12 +8,17 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 _Changes in the next release_
 
+### Added
+- Automatic configuration and setup through the [Unfolded Circle for Home Assistant component](https://github.com/JackJPowell/hass-unfoldedcircle). In cooperation with @albaintor and @JackJPowell, thanks! ([#60](https://github.com/unfoldedcircle/integration-home-assistant/pull/60)).  
+  - Please note that the autoconfiguration feature in the Home Assistant component is still under development. 
+### Fixed
+- Avoid initial failed Home Assistant login attempt with default HA server url and empty access token.
 
 ---
 
 ## v0.10.0 - 2024-08-24
 ### Added
-- Initial support for the UC unified integration in Home Assistant for optimized message communication. Contributed by @albaintor, thanks! ([#58](https://github.com/unfoldedcircle/integration-home-assistant/pull/58))
+- Initial support for the [Unfolded Circle for Home Assistant component](https://github.com/JackJPowell/hass-unfoldedcircle) for optimized message communication. Contributed by @albaintor, thanks! ([#58](https://github.com/unfoldedcircle/integration-home-assistant/pull/58))
 
 ### Changed
 - Update uc_api crate to latest 0.12.0 version.